# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Agent OS - AI Portable Computer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: agents
    description: Agent registration and management
  - name: commands
    description: Computer control commands
  - name: system
    description: System information and state
  - name: files
    description: Virtual filesystem operations
  - name: memory
    description: Agent memory and knowledge base
  - name: code
    description: Code execution engine
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List registered agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: registerAgent
      tags: [agents]
      summary: Register a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAgentRequest"
      responses:
        "201":
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"

  /agents/{agentId}:
    delete:
      operationId: unregisterAgent
      tags: [agents]
      summary: Unregister an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent unregistered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /commands:
    get:
      operationId: listCommands
      tags: [commands]
      summary: Get pending and recent commands
      responses:
        "200":
          description: Command list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"
    post:
      operationId: executeCommand
      tags: [commands]
      summary: Execute a computer control command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandRequest"
      responses:
        "200":
          description: Command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"

  /system/screenshot:
    get:
      operationId: getScreenshot
      tags: [system]
      summary: Get current screenshot of the desktop
      responses:
        "200":
          description: Screenshot data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenshotResponse"

  /system/state:
    get:
      operationId: getSystemState
      tags: [system]
      summary: Get current system state
      responses:
        "200":
          description: System state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemState"

  /system/apps:
    get:
      operationId: listApps
      tags: [system]
      summary: List available applications
      responses:
        "200":
          description: Application list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"

  /tasks:
    get:
      operationId: listTasks
      tags: [commands]
      summary: List all tasks
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [commands]
      summary: Create and start a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [commands]
      summary: Get task status
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files and directories
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VirtualFile"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file or directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileRequest"
      responses:
        "201":
          description: File created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualFile"

  /files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file content
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualFile"
    put:
      operationId: updateFile
      tags: [files]
      summary: Update file content
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFileRequest"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualFile"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete file or directory
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /memory:
    get:
      operationId: listMemories
      tags: [memory]
      summary: List agent memories
      parameters:
        - name: agentId
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Memory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memory]
      summary: Store a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryRequest"
      responses:
        "201":
          description: Memory stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"

  /memory/{memoryId}:
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete a memory
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /code/run:
    post:
      operationId: runCode
      tags: [code]
      summary: Execute code in a sandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCodeRequest"
      responses:
        "200":
          description: Code execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunCodeResult"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        endpoint:
          type: string
        capabilities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, busy]
        connectedAt:
          type: string
        lastSeen:
          type: string
      required:
        - id
        - name
        - capabilities
        - status
        - connectedAt

    RegisterAgentRequest:
      type: object
      properties:
        name:
          type: string
        endpoint:
          type: string
        capabilities:
          type: array
          items:
            type: string
      required:
        - name
        - capabilities

    CommandRequest:
      type: object
      properties:
        agentId:
          type: string
        type:
          type: string
          enum: [click, double_click, type, scroll, key, screenshot, open_app, close_app, terminal, navigate, drag]
        payload:
          type: object
          additionalProperties: true
      required:
        - type
        - payload

    Command:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        type:
          type: string
        payload:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pending, executing, completed, failed]
        createdAt:
          type: string
        completedAt:
          type: string
        result:
          type: object
          additionalProperties: true
      required:
        - id
        - type
        - payload
        - status
        - createdAt

    CommandResult:
      type: object
      properties:
        commandId:
          type: string
        success:
          type: boolean
        result:
          type: object
          additionalProperties: true
        error:
          type: string
      required:
        - commandId
        - success

    ScreenshotResponse:
      type: object
      properties:
        dataUrl:
          type: string
        timestamp:
          type: string
        width:
          type: number
        height:
          type: number
      required:
        - timestamp

    SystemState:
      type: object
      properties:
        runningApps:
          type: array
          items:
            type: string
        cpu:
          type: number
        memory:
          type: number
        uptime:
          type: number
        connectedAgents:
          type: number
        totalCommands:
          type: number
      required:
        - runningApps
        - cpu
        - memory
        - uptime
        - connectedAgents
        - totalCommands

    App:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        isOpen:
          type: boolean
      required:
        - id
        - name
        - icon
        - description
        - isOpen

    Task:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TaskStep"
        createdAt:
          type: string
        completedAt:
          type: string
        progress:
          type: number
      required:
        - id
        - name
        - status
        - steps
        - createdAt
        - progress

    TaskStep:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        commandType:
          type: string
        result:
          type: string
      required:
        - id
        - description
        - status

    CreateTaskRequest:
      type: object
      properties:
        agentId:
          type: string
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              commandType:
                type: string
              payload:
                type: object
                additionalProperties: true
            required:
              - description
              - commandType
      required:
        - name
        - steps

    VirtualFile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
        size:
          type: number
        mimeType:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - path
        - type
        - createdAt
        - updatedAt

    CreateFileRequest:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
        mimeType:
          type: string
      required:
        - name
        - path
        - type

    UpdateFileRequest:
      type: object
      properties:
        content:
          type: string
        name:
          type: string
      required: []

    Memory:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        type:
          type: string
          enum: [episodic, semantic, procedural, note]
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        importance:
          type: number
        createdAt:
          type: string
      required:
        - id
        - type
        - content
        - tags
        - importance
        - createdAt

    CreateMemoryRequest:
      type: object
      properties:
        agentId:
          type: string
        type:
          type: string
          enum: [episodic, semantic, procedural, note]
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        importance:
          type: number
      required:
        - type
        - content

    RunCodeRequest:
      type: object
      properties:
        language:
          type: string
          enum: [javascript, python, bash]
        code:
          type: string
        agentId:
          type: string
      required:
        - language
        - code

    RunCodeResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: number
        executionTime:
          type: number
        success:
          type: boolean
      required:
        - stdout
        - stderr
        - exitCode
        - executionTime
        - success

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error