# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent OS - AI Portable Computer API
 * OpenAPI spec version: 0.2.0
 */

export interface App {
  id: string;
  name: string;
  icon: string;
  description: string;
  isOpen: boolean;
}