# lib/api-zod/src/generated/types/commandRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent OS - AI Portable Computer API
 * OpenAPI spec version: 0.2.0
 */
import type { CommandRequestPayload } from "./commandRequestPayload";
import type { CommandRequestType } from "./commandRequestType";

export interface CommandRequest {
  agentId?: string;
  type: CommandRequestType;
  payload: CommandRequestPayload;
}