# lib/api-zod/src/generated/types/createFileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent OS - AI Portable Computer API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateFileRequestType } from "./createFileRequestType";

export interface CreateFileRequest {
  name: string;
  path: string;
  type: CreateFileRequestType;
  content?: string;
  mimeType?: string;
}