# lib/api-zod/src/generated/types/createMemoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent OS - AI Portable Computer API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateMemoryRequestType } from "./createMemoryRequestType";

export interface CreateMemoryRequest {
  agentId?: string;
  type: CreateMemoryRequestType;
  content: string;
  tags?: string[];
  importance?: number;
}