# lib/api-zod/src/generated/types/runCodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent OS - AI Portable Computer API
 * OpenAPI spec version: 0.2.0
 */
import type { RunCodeRequestLanguage } from "./runCodeRequestLanguage";

export interface RunCodeRequest {
  language: RunCodeRequestLanguage;
  code: string;
  agentId?: string;
}