# lib/api-zod/src/generated/types/updateFileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent OS - AI Portable Computer API
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateFileRequest {
  content?: string;
  name?: string;
}