# lib/api-zod/src/generated/types/virtualFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent OS - AI Portable Computer API
 * OpenAPI spec version: 0.2.0
 */
import type { VirtualFileType } from "./virtualFileType";

export interface VirtualFile {
  id: string;
  name: string;
  path: string;
  type: VirtualFileType;
  content?: string;
  size?: number;
  mimeType?: string;
  createdAt: string;
  updatedAt: string;
}